fix: address several findbugs/compiler warnings that appear in the console during builds

+++ uPortal-utils/uPortal-utils-core/src/main/java/org/apereo/portal/utils/cache/CacheManagementHelper.java
@@ -70,8 +70,8 @@ public class CacheManagementHelper {
                 cacheStatistics.misses = statistics.getCacheMisses();
                 cacheStatistics.size = statistics.getObjectCount();
                 cacheStatistics.maxSize =
-                        cacheConfiguration.getMaxElementsInMemory()
-                                + cacheConfiguration.getMaxElementsOnDisk();
+                        (long) cacheConfiguration.getMaxElementsInMemory()
+                                + (long) cacheConfiguration.getMaxElementsOnDisk();
 
                 allCacheStatistics.put(cacheName, cacheStatistics);
             }
